pyproject.toml
[tool.poetry]
name = "langchain-tutorials"
version = "0.1.0"
description = ""
authors = ["Arcadia822 <arcadia822@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "langchain_tutorials"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.7"
jupyter = "^1.0.0"
langchain = "^0.0.189"
pydantic = "^1.10.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
